% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - salmotools.R
\name{adv_calc_TGC}
\alias{adv_calc_TGC}
\title{Calculates Thermal Growth Coefficient TGC}
\usage{
adv_calc_TGC(df_fish, df_temp, period)
}
\description{
Calculates Thermal Growth Coefficient TGC
}
\keyword{internal}
