% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - genomatchomatico.R
\name{assign_closest_matches}
\alias{assign_closest_matches}
\title{assign_closest_matches}
\usage{
assign_closest_matches(
  df_samples,
  df_lookup,
  similarity_matrix,
  cutoff_similarity = 0.9,
  conflicts_resolve = T
)
}
\arguments{
\item{df_samples}{Data frame with all samples to be lookup up}

\item{df_lookup}{Data frame with samples to be looked up against}

\item{similarity_matrix}{If you already have made a similarity matrix (see function create_similarityMatrix()) you can refer to it here and save some time.}

\item{cutoff_similarity}{Any match with a similarity lower than this is set as NA}

\item{resolve_conflicts}{If conflicts (samples with the same match) should be resolved}
}
\description{
Takes a dataframe of genetic samples, and checks it against another frame of samples to find matching samples.
Both dataframes must be formated as:\n Rows=individuals/samples \n Column=SNP genotypes. Spelled as bases, e.g. "AB", "CT", "CC" etc. Must contain a column "ID" with a unique identifier for each row.
\n\n
Returns df_samples with an additional column, "ID_match", which is the closest match of that individual
\n\n
Needs to create a similarity matrix to operate. This is time consuming, and if you have already created one you can refer to it using the parameter similarity_matrix.
After creating the similarity matrix, the function saves the matrix as a csv file for use later.
}
