% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_pgs.R
\name{available_pgs}
\alias{available_pgs}
\title{Available PGS}
\usage{
available_pgs(
  geno_data = "MoBaPsychGen_v1-ec-eur-batch-basic-qc",
  pgs_directory = "//ess01/P471/data/durable/common/pgs_directory/pgs",
  pgs_software = "ldpred2",
  pgs_meta = genotools::pgs_metadata,
  within_fam = FALSE
)
}
\arguments{
\item{geno_data}{Which genotype data files do scores come from? Defaults to
"MoBaPsychGen_v1-ec-eur-batch-basic-qc"; previous versions also  available for
continuity (but not recommended for new projects): "98k-mobagenetics-ieu",
 "98k-ec-eur-fin-batch-basic-qc","hrv_rot1"}

\item{pgs_directory}{Where are the PGS stored locally? Defaults to
"//ess01/P471/data/durable/common/pgs_directory/pgs"}

\item{pgs_software}{With what software were PGS created? Defaults to "prsice2"}

\item{pgs_meta}{What R object holds the PGS metadata? Defaults to
genotools::pgs_metadata}

\item{within_fam}{Are scores for within-family analyses such as trio-PGS? If so,
you may prefer to use a score based on only very well imputed (INFO>0.95) variants (see
https://www.medrxiv.org/content/10.1101/2024.10.01.24314703v1.full-text for
context on this decision). Defaults to FALSE.}
}
\description{
\code{available_pgs} returns a dataframe with details of already available
polygenic scores (PGS)
}
\details{
Detailed description...
}
