% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - genomatchomatico.R
\name{calculate_distance_to_sample}
\alias{calculate_distance_to_sample}
\title{find closest match}
\usage{
calculate_distance_to_sample(table, sample, notSNP)
}
\arguments{
\item{sample}{the sample to find closest match to}

\item{lookup}{a dataframe of samples with SNP data (Each row one sample, each column a snp loki). First row must be ID info.}
}
\description{
Using the find_genetic_similarity function, takes a table of genetic samples (rows: samples, columns: snp genotypes) and calculates
}
