% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - base genotools.R
\name{clean_ID}
\alias{clean_ID}
\title{clean_ID}
\usage{
clean_ID(
  dataset,
  column,
  identifier = "",
  trailing_ident = F,
  numLength = 4,
  prefix = "",
  numeric = F,
  keepName = F
)
}
\arguments{
\item{dataset}{the dataset we're using. Each row has it's own ID}

\item{column}{the name of the column containing the ID's}

\item{identifier}{ID's need to be formated with a number and following expresio, e.g "34_individuals2019" where "_individuals2019" is the expression. Any entries not matching this format will be removed.}

\item{trailing_ident}{Wether the expression if before (F) or after (T)of the ID number}

\item{numLength}{if you want leading zeroes, use this parameter to specify the length of the number, e.g "8" for 00000342}

\item{prefix}{if you want a prefix in the new cleaned ID. Ex: "individuals2019_" will give you "individuals2019_0034"s}

\item{keepName}{T: keeps the old column name, F: renames it to "ID", any other string: Renames the column to this string}
}
\description{
For cleaning "dirty" ID's and removing rows that doesn't match our ID-pattern
}
