% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - base genotools.R
\name{determineSex}
\alias{determineSex}
\title{determineSex
For determining sex based on SDY}
\usage{
determineSex(dataframe, column, cutoff)
}
\description{
determineSex
For determining sex based on SDY
}
