% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_pgs.R
\name{fetch_pgs}
\alias{fetch_pgs}
\title{Fetch PGS from directory}
\usage{
fetch_pgs(
  pgs_list = NULL,
  thresholds = c(5e-08, 5e-07, 5e-06, 5e-05, 5e-04, 0.001, 0.01, 0.005, 0.1, 0.5, 1),
  threshold_range = FALSE,
  geno_data = "MoBaPsychGen_v1-ec-eur-batch-basic-qc",
  pgs_directory = "//ess01/P471/data/durable/common/pgs_directory/pgs/ldpred2/no_filter",
  pgs_software = "ldpred2",
  pgs_meta = genotools::pgs_metadata,
  maf = "0.05",
  clump = "500_1_0.25",
  within_fam = FALSE,
  zip_file = TRUE
)
}
\arguments{
\item{pgs_list}{a string or vector of strings from the Pheno_shortname column of available_pgs()}

\item{thresholds}{(ignored unless using a clumping and thresholding approach to PGS creation)
a vector of numeric values corresponding to the p-value thresholds you want
or a single value between 0-1 to return all thresholds lower than this value by using the
threshold_range = TRUE option; defaults to c(5e-08, 5e-07, 5e-06, 5e-05, 5e-04, 0.001,
0.01, 0.005, 0.1, 0.5, 1)}

\item{threshold_range}{(ignored unless using a clumping and thresholding approach to PGS creation)
use TRUE with a single value in thresholds to get a range; defaults to FALSE}

\item{geno_data}{Which genotype data files do scores come from? Defaults to
"MoBaPsychGen_v1-ec-eur-batch-basic-qc"; previous versions also available for continuity (but
 not recommended for new projects): "98k-mobagenetics-ieu","98k-ec-eur-fin-batch-basic-qc","hrv_rot1"}

\item{pgs_directory}{Where are the PGS stored locally? Defaults to
"//ess01/P471/data/durable/common/pgs_directory/pgs/ldpred2/no_filter"}

\item{pgs_software}{With what software were PGS created? Defaults to "ldpred2"}

\item{pgs_meta}{What R object holds the PGS metadata? Defaults to genotools::pgs_metadata}

\item{maf}{Used for retrieving scores made with PRSice2 based on the clumping and
thresholding method; must be a string from pgs directory tree e.g. "0.05"; see all possibilities
 using \code{stringr::str_remove_all(unique(available_pgs()$maf), "maf")}}

\item{clump}{Used for retrieving scores made with PRSice2 based on the clumping and
thresholding method; must be a string from pgs directory tree e.g. "500_1_0.25"; see all possibilities
 using \code{stringr::str_remove_all(unique(available_pgs()$clump), "clump")}}

\item{within_fam}{Are scores for within-family analyses such as trio-PGS? If so,
you may prefer to use a score based on only very well imputed (INFO>0.95) variants (see
https://www.medrxiv.org/content/10.1101/2024.10.01.24314703v1.full-text for
context on this decision). Defaults to FALSE.}

\item{zip_file}{is the file for the PGS saved as .gz, defualt is = TRUE}
}
\description{
\code{fetch_pgs} returns a dataframe with requested polygenic scores (PGS)
}
\details{
Detailed description...
}
