% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pgs_info.R
\name{get_pgs_log}
\alias{get_pgs_log}
\alias{get_pgs_snps}
\alias{get_pgs_nsnps}
\title{Get info about PGS from directory}
\usage{
get_pgs_log(
  pgs_list = NULL,
  thresholds = c(5e-08, 5e-07, 5e-06, 5e-05, 5e-04, 0.001, 0.01, 0.005, 0.1, 0.5, 1),
  geno_data = "MoBaPsychGen_v1-ec-eur-batch-basic-qc",
  pgs_directory = "//ess01/P471/data/durable/common/pgs_directory/pgs",
  pgs_software = "prsice2",
  maf = "0.01",
  clump = "250_1_0.1"
)

get_pgs_snps(
  pgs_list = NULL,
  geno_data = "MoBaPsychGen_v1-ec-eur-batch-basic-qc",
  pgs_directory = "//ess01/P471/data/durable/common/pgs_directory/pgs",
  pgs_software = "prsice2",
  maf = "0.01",
  clump = "250_1_0.10"
)

get_pgs_nsnps(
  pgs_list = NULL,
  thresholds = c(5e-08, 5e-07, 5e-06, 5e-05, 5e-04, 0.001, 0.01, 0.005, 0.1, 0.5, 1),
  geno_data = "MoBaPsychGen_v1-ec-eur-batch-basic-qc",
  pgs_directory = "//ess01/P471/data/durable/common/pgs_directory/pgs",
  pgs_software = "prsice2",
  maf = "0.01",
  clump = "250_1_0.10"
)
}
\arguments{
\item{pgs_list}{a string or vector of strings from the Pheno_shortname column of available_pgs()}

\item{thresholds}{a vector of numeric values corresponding to the p-value thresholds you want}

\item{geno_data}{Which genotype data files do scores come from? Defaults to
"MoBaPsychGen_v1-ec-eur-batch-basic-qc"; also supports "98k-mobagenetics-ieu",
"98k-ec-eur-fin-batch-basic-qc", "hrv_rot1" for legacy reasons - these should
not be needed for new analyses}

\item{pgs_directory}{Where are the PGS stored locally? Defaults to
"//ess01/P471/data/durable/common/pgs_directory/pgs"}

\item{pgs_software}{With what software were PGS created? Defaults to "prsice2"}

\item{maf}{must be a string from pgs directory tree e.g. "0.05"; see all possibilities
using \code{stringr::str_remove_all(unique(available_pgs()$maf), "maf")}}

\item{clump}{must be a string from pgs directory tree e.g. "500_1_0.25"; see all possibilities
using \code{stringr::str_remove_all(unique(available_pgs()$clump), "clump")}}
}
\description{
\code{get_pgs_log} copies the log files for the PGS creation into a specified directory
\code{get_pgs_snps} get lists of all SNPs available after clumping for any set of PGS
\code{pgs_nspns_thresh} get tables with number of SNPs at each thresholds for any set of PGS
}
\details{
Detailed description...
}
