% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pgs.R
\name{make_pgs}
\alias{make_pgs}
\title{Make and run ob script for polygenic score creation}
\usage{
make_pgs(
  user,
  host = "p471-hpc-01.tsd.usit.no",
  software = "ldpred2",
  shortname,
  account = "p471",
  cpu_time = "05:00:00",
  memory = "8G",
  cores = 32,
  lib = "//ess/p471/data/durable/common/rlibs/4-2-1",
  sumstats_dir = "//ess/p471/data/no-backup/gwas_sumstats/",
  outputs_dir = NA,
  sumstats_filename = NA,
  genotype_dir = "//ess/p471/data/durable/data/genetic/MoBaPsychGen_v1/",
  genotype_data = "genoHapMap3plus_N200k.rds",
  ldref_dir = "//ess/p471/data/durable/data/genetic/ldref/",
  case_control = TRUE,
  A1 = NA,
  A2 = NA,
  stat = NA,
  stat_type = NA,
  pvalue = NA,
  SNP = NA,
  CHR = NA,
  BP = NA,
  SE = NA,
  N = NA,
  info = NA,
  MAF = NA,
  maf_threshold = 0.01,
  source = NA,
  pmid = NA,
  sd_maf = FALSE,
  pgs_meta = genotools::pgs_metadata,
  session = NULL,
  submit_ssh = TRUE
)
}
\arguments{
\item{user}{username for cluster}

\item{host}{hostname for cluster - defaults to p471 submit node}

\item{software}{which software to used to make the PGS? Currently "ldpred2" (default) or "prsice2"}

\item{shortname}{shortname for the PGS - must be distinct from others in available_pgs()$Pheno_shortname}

\item{account}{the colossus/SLURM account from which CPU useage will be deducted for making the scores;
in p471, this can be "p471" (the default) or "p471_tsd" - use 'qsumm' command in bash to see availability}

\item{cpu_time}{how long is required for the job on Colossus (defaults to 05:00:00)}

\item{memory}{how much memory per core is required for the job on Colossus (defaults to 8GB)}

\item{cores}{how many cores should be used? Defaults to 32}

\item{lib}{(optional) add a path to a common R library with required packages installed}

\item{sumstats_dir}{directory for sumstats, suggest leaving at default if working in p471 - NB, this filepath
will be used by the HPC on linux, so should be specified accordingly ("ess" not "ess01")}

\item{outputs_dir}{directory for outputs; NB, this filepath
will be used by the HPC on linux, so should be specified accordingly ("ess" not "ess01")}

\item{sumstats_filename}{sumstats filename, including extension}

\item{genotype_dir}{location of genotype data on cluster; defaults for mobapsychgen_v1}

\item{genotype_data}{name of genotype data on cluster, formatted as an RDS file containing the subset
of the MoBa genetic data that overlaps with HapMap3 SNPs...; defaults set for mobapsychgen_v1}

\item{ldref_dir}{location of the HAPMAP3+ and ld blocks needed by ldpred2}

\item{case_control}{whether the GWAS is case control - will try to pull from existing metadata if left blank; defaults to TRUE}

\item{A1}{effect allele column name - will try to pull from existing metadata if left blank}

\item{A2}{other allele column - will try to pull from existing metadata if left blank}

\item{stat}{stat column - will try to pull from existing metadata if left blank}

\item{stat_type}{whether the stat is beta or OR- will try to pull from existing metadata if left blank}

\item{pvalue}{pval column - will try to pull from existing metadata if left blank}

\item{SNP}{snp id column - will try to pull from existing metadata if left blank}

\item{CHR}{CHR column - will try to pull from existing metadata if left blank}

\item{BP}{bp column - will try to pull from existing metadata if left blank}

\item{SE}{standard error column - will try to pull from existing metadata if left blank}

\item{N}{either, if trait is continuous a) a column name with SNP-wise Ns or b) a value
representing the overall N, or - if binary - c) a comma-separated pair of column names for SNP-wise
N-cases and N-controls or d) a vector of values representing the N-cases and N-controls}

\item{info}{(optional) info score column name for filtering}

\item{MAF}{(optional) maf column name for filtering - NB can also provide EAF here}

\item{maf_threshold}{minor allele frequency threshold - defaults to 0.01}

\item{source}{Where do the summary stats come from? (ideally a url) - either this or pmid
must be non-missing - will try to pull from existing metadata if left blank}

\item{pmid}{What publication do the summary stats relate to? - either this or source must be
non-missing - will try to pull from existing metadata if left blank}

\item{sd_maf}{calculate sd from an allele freq in summary stats - default is FALSE}

\item{pgs_meta}{What R object holds the PGS metadata? Defaults to genotools::pgs_metadata}

\item{session}{(optional) provide an active SSH session to avoid repeated login prompts if
mapping or looping this function}

\item{submit_ssh}{true/false to submit the job to the cluster. Default  is true. If false dir, .pipl file, and .sh file are created and can be edited and submitted manually.}
}
\description{
\code{make_pgs} makes and submits a submission script for
running on Colossus to make polygenic scores
}
\details{
Detailed description...
}
