% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pgs_batch.R
\name{make_pgs_batch}
\alias{make_pgs_batch}
\title{Make a batch of polygenic scores using the \code{make_pgs} function}
\usage{
make_pgs_batch(
  inputs,
  user,
  host = "p471-hpc-01.tsd.usit.no",
  software = "ldpred2",
  shortname,
  account = "p471",
  cpu_time = "05:00:00",
  memory = "8G",
  cores = 32,
  lib = "//ess/p471/data/durable/common/rlibs/4-2-1",
  sumstats_dir = "//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
  outputs_dir = "//ess/p471/data/durable/common/pgs_directory",
  genotype_dir = "//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
  genotype_data = "genoHapMap3plus_N200k.rds",
  ldref_dir = "//ess/p471/data/durable/data/genetic/ldref/",
  maf_threshold = 0.01,
  ...
)
}
\arguments{
\item{inputs}{a dataframe containing metadata to allow PGS to be created. The
formatting should should match genotools::pgs_metadata}

\item{user}{username for cluster}

\item{host}{hostname for cluster - defaults to p471 submit node}

\item{software}{which software to used to make the PGS? Currently "ldpred2" (default) or "prsice2"}

\item{shortname}{shortname for the PGS - must be distinct from others in available_pgs()$Pheno_shortname}

\item{account}{the colossus/SLURM account from which CPU useage will be deducted for making the scores;
in p471, this can be "p471" (the default) or "p471_tsd" - use 'qsumm' command in bash to see availability}

\item{cpu_time}{how long is required for the job on Colossus (defaults to 05:00:00)}

\item{memory}{how much memory per core is required for the job on Colossus (defaults to 8GB)}

\item{cores}{how many cores should be used? Defaults to 32}

\item{lib}{(optional) add a path to a common R library with required packages installed}

\item{sumstats_dir}{directory for sumstats, suggest leaving at default if working in p471 - NB, this filepath
will be used by the HPC on linux, so should be specified accordingly ("ess" not "ess01")}

\item{outputs_dir}{directory for outputs; NB, this filepath
will be used by the HPC on linux, so should be specified accordingly ("ess" not "ess01")}

\item{genotype_dir}{location of genotype data on cluster; defaults for mobapsychgen_v1}

\item{genotype_data}{name of genotype data on cluster, formatted as an RDS file containing the subset
of the MoBa genetic data that overlaps with HapMap3 SNPs...; defaults set for mobapsychgen_v1}

\item{ldref_dir}{location of the HAPMAP3+ and ld blocks needed by ldpred2}

\item{...}{arguments to pass to internal functions, see ?make_pgs}
}
\description{
\code{make_pgs_batch} is used to make a batch of polygenic scores based
on a data.frame of inputs (see \code{help(make_pgs)} for main inputs)
}
\details{
\code{make_pgs} creates and runs a job script to make a single PGS based on
 user-supplied arguments containing information about column names etc.
 \code{make_pgs_batch} is a wrapper script to allow a batch of PGS to be made at
 the same time, woth
}
