% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgs_pca.R
\name{pgs_pca}
\alias{pgs_pca}
\title{PGS PCA}
\usage{
pgs_pca(dat, indid, pgs_var_stem, return = "pcs")
}
\arguments{
\item{indid}{individual level id variable}

\item{pgs_var_stem}{string(s) for the stem that is in common for all variables}

\item{return}{some combination of "pcs", "r2s", and "loadings"
representing PGS for a given phenotype at some p-value threshold}

\item{geno_dat}{data frame with polgenic scores (probably from \code{process_pgs})}
}
\description{
\code{pgs_pca} implements the PGS-PCA method to return the first principle
component from PCA of a set of scores at a range of thresholds made using
a clumping and thresholding method such as prsice2 (not applicable for
scores from ldpred2)
}
\details{
See https://doi.org/10.1002/gepi.22339
}
