% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pgs.R
\name{process_pgs}
\alias{process_pgs}
\title{Process PGS fetched from directory}
\usage{
process_pgs(
  fetched_pgs_df,
  regress_pgs = TRUE,
  geno_data = "MoBaPsychGen_v1-ec-eur-batch-basic-qc",
  covs_dir = "//ess01/P471/data/durable/data/genetic/MoBaPsychGen_v1",
  return_covs = TRUE,
  analysis_type = NULL
)
}
\arguments{
\item{fetched_pgs_df}{input should be data.frame results of fetch_pgs}

\item{regress_pgs}{do you want PGS regressed on batch and 20PCs? Defaults to TRUE}

\item{geno_data}{Which genotype data files do scores come from? Defaults to
"MoBaPsychGen_v1-ec-eur-batch-basic-qc"; also supports "98k-mobagenetics-ieu",
"98k-ec-eur-fin-batch-basic-qc", "hrv_rot1" for legacy reasons - these should
not be needed for new analyses}

\item{covs_dir}{Where is the covariate file for the geno_data? Defaults to
"//ess01/P471/data/durable/data/genetic/MoBaPsychGen_v1" for p471}

\item{return_covs}{Do you want batch and PCS returned? Defaults to TRUE}

\item{analysis_type}{Not yet implemented (for exclusion flag list vars)}
}
\description{
\code{fetch_pgs} takes raw PGS and joins to phenotypic data IDs, whilst also
regressing on key covariates and/or returning these covariates for inclusion in
analyses.
}
\details{
Detailed description...
}
