% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - genomatchomatico.R
\name{resolve_conflicts}
\alias{resolve_conflicts}
\title{resolve_conflicts}
\usage{
resolve_conflicts(df_samples, df_similarities)
}
\description{
For use together with assign_closest_matches
}
