% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - run_snppit.R
\name{run_snppit}
\alias{run_snppit}
\title{run_snppt}
\usage{
run_snppit(
  data_offspring,
  data_parents,
  projectName = "project1",
  overWrite = F
)
}
\arguments{
\item{data_offspring}{Dataset contaning offspring ID's and SNP genotypes (see above)}

\item{data_parents}{Dataset containing parent ID's, sex, population, and SNP genotypes (see above)}

\item{projectName}{Optional. A name that will be used for files created during the proces}
}
\description{
Ver 0.1
}
\details{
Function that simplifies using SNPPT.exe to calculate parentage based on SNP data
Requires that snppt.exe is installed in the working directory!

Takes two datasets, one for offspring and one for parents. \cr
Both datasets must be formated as folows:
\itemize{
\item Individuals as rows
\item SNP data on columns, one column pr SNP
\item SNP genotype written numerically: 1=homozygous for allele a, 2=heterozygous, 3=homozygous for allele b
\item One column named "ID" that contains some ID for each individual,
\item Only for parents: One column named "population" that's either "M" or "F"
\item Only for parents: One column named "population" that gives that individual's population
}

Important!: Any column that is not ID, sex (parents) or population (parents) will be interpreted as an SNP!!!
}
\examples{
run_snppt(offspring, parents, "Project_oct2019")
}
