% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrelate.R
\name{unrelate}
\alias{unrelate}
\title{Get an unrelated sub-sample of MoBa data based on your analysis type}
\usage{
unrelate(
  input_df = NULL,
  unrelated = "trios",
  complete = FALSE,
  phenotypes = NULL,
  geno_data = "MoBaPsychGen_v1-ec-eur-batch-basic-qc",
  covs_dir = "//ess01/P471/data/durable/data/genetic/MoBaPsychGen_v1",
  keep_flag = FALSE
)
}
\arguments{
\item{input_df}{a data.frame with at least i) preg_id + BARN_NR;
ii) FID and IID; or (if analysis only includes parents) iii) m_id and/or f_id}

\item{unrelated}{a string describing who must be unrelated to one another in
the output; can be "trios" (for unrelated trios), "nuclear" (for unrelated
nucleary families - i.e., trios + other children) "children" (for unrelated
children, excluding even all-but-one child per nuclear family), "sibs" (for
unrelated children outside nuclear family sib pairs), "mothers", "fathers",
"parents","mcduos" (for unrelated sets of mother-child duos), or "fcduos"
(for unrelated sets of father-child duos)}

\item{complete}{should only complete sets of the specified unrelated
groupings be returned? Defaults to FALSE, meaning that incomplete
groupings can also be returned (i.e., mother child duos when unrelated = "trios" -
as long as they are not related to an included complete trio)}

\item{phenotypes}{column names from 'input_df' containing phenotypes you want to
prioritise data availability for when excluding related individuals}

\item{geno_data}{Which genotype data files do scores come from? Defaults to
"MoBaPsychGen_v1-ec-eur-batch-basic-qc"; legacy genotype data files not supported
by this function (cf. the _pgs functions)}

\item{covs_dir}{Where is the covariate file for the geno_data? Defaults to
"//ess01/P471/data/durable/data/genetic/MoBaPsychGen_v1" for p471}

\item{keep_flag}{Keep related individuals and add a flag variable? Defaults
to FALSE - i.e., remove them from the final dataset}
}
\description{
\code{unrelate} takes a MoBa dataset with IDs (genetic, phenotypic, or both),
and - based on the kind of relatedness you declare as allowed in your analysis -
removes individuals to limit additional relatedness while prioritising
genotype and phenotype data availability
}
\details{
Detailed description...
}
