% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgs_search.R
\name{variable_search}
\alias{variable_search}
\title{PGS search}
\usage{
variable_search(dataframe, string, where = "anywhere")
}
\arguments{
\item{dataframe}{Where to look for variables -should be a data frame returned
by a call to \code{available_pgs}}

\item{string}{a vector of strings to search for}

\item{where}{if you know the column name in which the string should be found,
you can supply it here; the default is to search in all columns ("anywhere")}
}
\description{
\code{pgs_search} is a helper to find PGS that are available by
a string anywwere in the variable name or description
}
\details{
Detailed description..
}
