% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - salmotools.R
\name{adv_calc_TGC}
\alias{adv_calc_TGC}
\title{improved TGC calculator. Returns a vector}
\usage{
adv_calc_TGC(df_fish, df_temp, period)
}
\description{
improved TGC calculator. Returns a vector
}
\keyword{internal}
